refactor code, move express middleware to a seperate file
